+++ Prowl.Runtime/GameObject/LayerMask.cs
@@ -18,9 +18,9 @@ public struct LayerMask
 
     public void Clear() => mask = 0;
 
-    public bool HasLayer(byte index) => (mask & (1 << index)) == (1 << index);
-    public void SetLayer(byte index) => mask |= 1u << index;
-    public void RemoveLayer(byte index) => mask &= ~(1u << index);
+    public bool HasLayer(int index) => (mask & (1 << index)) == (1 << index);
+    public void SetLayer(int index) => mask |= 1u << index;
+    public void RemoveLayer(int index) => mask &= ~(1u << index);
     public static LayerMask operator |(LayerMask mask1, LayerMask mask2) => new() { mask = mask1.mask | mask2.mask };
     public static LayerMask operator &(LayerMask mask1, LayerMask mask2) => new() { mask = mask1.mask & mask2.mask };
     public override bool Equals(object? obj)
@@ -31,6 +31,6 @@ public struct LayerMask
     }
     public override int GetHashCode() => mask.GetHashCode();
 
-    public static string LayerToName(byte index) => TagLayerManager.GetLayer(index);
-    public static byte NameToLayer(string name) => TagLayerManager.GetLayerIndex(name);
+    public static string LayerToName(int index) => TagLayerManager.GetLayer(index);
+    public static int NameToLayer(string name) => TagLayerManager.GetLayerIndex(name);
 }
