+++ Prowl.Runtime/Rendering/RenderPipeline/IRenderable.cs
@@ -9,7 +9,7 @@ public interface IRenderable
 {
     public Material GetMaterial();
 
-    public byte GetLayer();
+    public int GetLayer();
 
     public void GetRenderingData(out PropertyState properties, out IGeometryDrawData drawData, out Matrix4x4 model);
